refactor: add communication module in infra and add node mailer
