+++ src/infra/communication/index.ts
@@ -1,0 +1,1 @@
+export * from './node-mailer-adapter'
