+++ src/main/factories/controllers/login-controller-factory.ts
@@ -1,13 +1,11 @@
 import { type Controller } from '@/application/protocols'
-import { LoginController } from '@/application/controllers/login'
-import { EmailValidatorAdapter } from '@/application/validation/validators'
+import { LoginController } from '@/application/controllers'
+import { EmailValidatorAdapter } from '@/application/validation'
 import { DbAuthentication } from '@/data/use-cases'
-import { BcryptAdapter } from '@/infra/cryptography/bcrypt-adapter'
-import { JwtAdapter } from '@/infra/cryptography/jwt-adapter'
-import { GuardianAccountRepository } from '@/infra/repos/postgresql/guardian-account-repository'
-import { LoggerPgRepository } from '@/infra/repos/postgresql/logger-repository'
-import env from '@/main/config/env'
+import { BcryptAdapter, JwtAdapter } from '@/infra/cryptography'
+import { GuardianAccountRepository, LoggerPgRepository } from '@/infra/repos/postgresql'
 import { LoggerControllerDecorator } from '@/main/decorators'
+import env from '@/main/config/env'
 
 export const makeLoginController = (): Controller => {
   const salt = Number(env.salt)
