+++ tests/src/infra/node-mailer-adapter.spec.ts
@@ -1,5 +1,5 @@
 import { type EmailService } from '@/data/protocols'
-import { NodeMailerAdapter } from '@/infra/node-mailer-adapter'
+import { NodeMailerAdapter } from '@/infra/communication'
 
 jest.mock('nodemailer', () => {
   return {
